backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
httpx>=0.27,<1.0
supabase>=2.4,<3.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
